requirements.txt
click==7.1.2
joblib==0.17.0
numpy==1.19.4
opencv-python==3.4.8.29
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
